pyproject.toml
[project]
name = "foi-rag"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "asyncpg",
    "openai",
    "anthropic",
    "pymupdf",
    "tiktoken",
    "python-dotenv",
    "pandas",
    "requests",
]
